Move ViewTestMixin to core
